Revert "Sum method bug #22"

+++ src/lt/bit/pizzeria/Order.java
@@ -26,6 +26,6 @@ class Order {
     @Override
     public String toString() {
         return name + " x " + quantity +
-                ", price: " + (price * quantity) + " eur";
+                ", price: " + (double) (price * quantity) + " eur";
     }
 }
